Refactor according to feedback
