+++ test/test_C_First_Order.py
@@ -1,0 +1,23 @@
+import unittest
+
+from deformation_measurement import C_First_Order
+
+class Test_C_First_Order(unittest.TestCase):
+	def test_define_deformed_subset(self):
+		#TODO: use the updated class version
+		print("needs class version")
+		'''
+		i, j, I_matrix, J_matrix, N, I, J, X, Y = define_deformed_subset(11, 20, 20, 0, 0, 0, 0, 0, 0)
+		
+		print(i)
+		print(j)
+		print(I_matrix)
+		print(J_matrix)
+		print(N)
+		print(I)
+		print(I.shape)
+		print(J)
+		print(J.shape)
+		print(X)
+		print(Y)
+		'''
